outsourced NPC class
